requirements.txt
python-dotenv
requests
beautifulsoup4
duckduckgo_search
python-telegram-bot==20.7